Move word chart to first page
