+++ website/src/lib/components/index.ts
@@ -5,7 +5,6 @@ export { default as Verse1 } from "./verse1/verse1.svelte";
 export { default as Verse2 } from "./verse2/verse2.svelte";
 export { default as Verse3 } from "./verse3/verse3.svelte";
 export { default as Verse4 } from "./verse4/verse4.svelte";
-export { default as Verse5 } from "./test/verse-test.svelte";
 export { default as Parallax } from "./parallax.svelte";
 export { default as Nav } from "./nav.svelte";
 
@@ -14,4 +13,3 @@ export { verse1Keyframes } from "./verse1/keyframe.svelte";
 export { verse2Keyframes } from "./verse2/keyframe";
 export { verse3Keyframes } from "./verse3/keyframe";
 export { verse4Keyframes } from "./verse4/keyframe";
-export { verse5Keyframes } from "./test/keyframe.svelte";
