+++ website/src/lib/components/verse1/word_chart.svelte
