+++ website/src/lib/keyframe.svelte.ts
@@ -3,8 +3,7 @@ import {
     verse1Keyframes,
     verse2Keyframes,
     verse3Keyframes,
-    verse4Keyframes,
-    verse5Keyframes
+    verse4Keyframes
 } from "./components";
 import { updateScroll } from "./scroll-override";
 
@@ -28,8 +27,7 @@ export const keyframes: Keyframe[] = [
     ...verse1Keyframes,
     ...verse2Keyframes,
     ...verse3Keyframes,
-    ...verse4Keyframes,
-    ...verse5Keyframes
+    ...verse4Keyframes
 ];
 
 /**
